% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{MOEAutils}
\alias{MOEAutils}
\title{Utilities for Multiobjective Otimization}
\description{
Utilities for Multiobjective Otimization
}
\details{
Utilities for multiobjective otimization with evolutionary algorithms

\tabular{ll}{
Package: \tab MOEAutils\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2020-08-24\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Aljosa Vodopija
}
